optimize BubbleSortView
